requirements.txt
requests~=2.31.0
pycryptodome==3.17

pytz~=2023.3.post1
flask~=3.0.0
cryptography~=42.0.7
PyJWT~=2.8.0
hyper~=0.7.0
httpx~=0.27.0